src/requirements.txt
pandas>=2.0.0
python-dateutil>=2.8.2
numpy>=1.24.0
pyarrow>=12.0.0
psycopg2-binary>=2.9.5
tqdm>=4.65.0
chardet>=5.1.0
psutil>=5.9.0
sqlalchemy>=2.0.0
PyQt5>=5.15.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
sqlparse
